# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Intelligence Pipeline API",
    "description": "Internal API for job collection pipeline management, dashboard metrics, source configuration, and event tracking.",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/jobs": {
      "get": {
        "operationId": "listJobs",
        "summary": "List jobs",
        "tags": [
          "Jobs"
        ],
        "description": "Returns a paginated, filterable list of jobs.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "source",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "stack",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100
            },
            "in": "query",
            "name": "minScore",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "remoteType",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "seniority",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "lastSeen_desc",
                "score_desc",
                "discovered_desc",
                "company_asc",
                "title_asc"
              ],
              "default": "lastSeen_desc"
            },
            "in": "query",
            "name": "sort",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string"
                          },
                          "sourceId": {
                            "type": "string"
                          },
                          "collector": {
                            "type": "string"
                          },
                          "discoveredVia": {
                            "type": "string"
                          },
                          "sourceUrl": {
                            "type": "string"
                          },
                          "canonicalUrl": {
                            "type": "string",
                            "nullable": true
                          },
                          "title": {
                            "type": "string"
                          },
                          "companyName": {
                            "type": "string"
                          },
                          "location": {
                            "type": "string",
                            "nullable": true
                          },
                          "remoteType": {
                            "type": "string",
                            "nullable": true
                          },
                          "salaryMin": {
                            "type": "number",
                            "nullable": true
                          },
                          "salaryMax": {
                            "type": "number",
                            "nullable": true
                          },
                          "currency": {
                            "type": "string",
                            "nullable": true
                          },
                          "seniority": {
                            "type": "string",
                            "nullable": true
                          },
                          "stackTags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "geoRestrictions": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "score": {
                            "type": "number"
                          },
                          "sourceTrustScore": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string"
                          },
                          "notifiedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "matchReasons": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "riskFlags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "recommendedAction": {
                            "type": "string",
                            "nullable": true
                          },
                          "postedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "discoveredAt": {
                            "type": "string"
                          },
                          "lastSeenAt": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "source",
                          "sourceId",
                          "collector",
                          "discoveredVia",
                          "sourceUrl",
                          "canonicalUrl",
                          "title",
                          "companyName",
                          "location",
                          "remoteType",
                          "salaryMin",
                          "salaryMax",
                          "currency",
                          "seniority",
                          "stackTags",
                          "geoRestrictions",
                          "score",
                          "sourceTrustScore",
                          "status",
                          "notifiedAt",
                          "matchReasons",
                          "riskFlags",
                          "recommendedAction",
                          "postedAt",
                          "discoveredAt",
                          "lastSeenAt",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "page",
                        "pageSize",
                        "total",
                        "totalPages",
                        "hasNextPage"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "operationId": "getJobDetail",
        "summary": "Get job details",
        "tags": [
          "Jobs"
        ],
        "description": "Returns detailed information about a specific job including its events.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string"
                    },
                    "sourceId": {
                      "type": "string"
                    },
                    "collector": {
                      "type": "string"
                    },
                    "discoveredVia": {
                      "type": "string"
                    },
                    "sourceUrl": {
                      "type": "string"
                    },
                    "canonicalUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "title": {
                      "type": "string"
                    },
                    "companyName": {
                      "type": "string"
                    },
                    "location": {
                      "type": "string",
                      "nullable": true
                    },
                    "remoteType": {
                      "type": "string",
                      "nullable": true
                    },
                    "salaryMin": {
                      "type": "number",
                      "nullable": true
                    },
                    "salaryMax": {
                      "type": "number",
                      "nullable": true
                    },
                    "currency": {
                      "type": "string",
                      "nullable": true
                    },
                    "seniority": {
                      "type": "string",
                      "nullable": true
                    },
                    "stackTags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "geoRestrictions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "score": {
                      "type": "number"
                    },
                    "sourceTrustScore": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    },
                    "notifiedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "matchReasons": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "riskFlags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "recommendedAction": {
                      "type": "string",
                      "nullable": true
                    },
                    "postedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "discoveredAt": {
                      "type": "string"
                    },
                    "lastSeenAt": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "normalizedTitle": {
                      "type": "string",
                      "nullable": true
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "requirements": {
                      "type": "string",
                      "nullable": true
                    },
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "jobId": {
                            "type": "string"
                          },
                          "eventType": {
                            "type": "string"
                          },
                          "metadata": {
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "jobId",
                          "eventType",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "source",
                    "sourceId",
                    "collector",
                    "discoveredVia",
                    "sourceUrl",
                    "canonicalUrl",
                    "title",
                    "companyName",
                    "location",
                    "remoteType",
                    "salaryMin",
                    "salaryMax",
                    "currency",
                    "seniority",
                    "stackTags",
                    "geoRestrictions",
                    "score",
                    "sourceTrustScore",
                    "status",
                    "notifiedAt",
                    "matchReasons",
                    "riskFlags",
                    "recommendedAction",
                    "postedAt",
                    "discoveredAt",
                    "lastSeenAt",
                    "createdAt",
                    "updatedAt",
                    "normalizedTitle",
                    "description",
                    "requirements",
                    "events"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{id}/status": {
      "patch": {
        "operationId": "updateJobStatus",
        "summary": "Update job status",
        "tags": [
          "Jobs"
        ],
        "description": "Updates a job status to a user-controlled status value.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "new",
                      "saved",
                      "discarded",
                      "applied",
                      "ignored",
                      "interviewing",
                      "offer",
                      "ghosted"
                    ]
                  }
                },
                "required": [
                  "status"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "source": {
                      "type": "string"
                    },
                    "sourceId": {
                      "type": "string"
                    },
                    "collector": {
                      "type": "string"
                    },
                    "discoveredVia": {
                      "type": "string"
                    },
                    "sourceUrl": {
                      "type": "string"
                    },
                    "canonicalUrl": {
                      "type": "string",
                      "nullable": true
                    },
                    "title": {
                      "type": "string"
                    },
                    "companyName": {
                      "type": "string"
                    },
                    "location": {
                      "type": "string",
                      "nullable": true
                    },
                    "remoteType": {
                      "type": "string",
                      "nullable": true
                    },
                    "salaryMin": {
                      "type": "number",
                      "nullable": true
                    },
                    "salaryMax": {
                      "type": "number",
                      "nullable": true
                    },
                    "currency": {
                      "type": "string",
                      "nullable": true
                    },
                    "seniority": {
                      "type": "string",
                      "nullable": true
                    },
                    "stackTags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "geoRestrictions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "score": {
                      "type": "number"
                    },
                    "sourceTrustScore": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    },
                    "notifiedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "matchReasons": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "riskFlags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "recommendedAction": {
                      "type": "string",
                      "nullable": true
                    },
                    "postedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "discoveredAt": {
                      "type": "string"
                    },
                    "lastSeenAt": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "source",
                    "sourceId",
                    "collector",
                    "discoveredVia",
                    "sourceUrl",
                    "canonicalUrl",
                    "title",
                    "companyName",
                    "location",
                    "remoteType",
                    "salaryMin",
                    "salaryMax",
                    "currency",
                    "seniority",
                    "stackTags",
                    "geoRestrictions",
                    "score",
                    "sourceTrustScore",
                    "status",
                    "notifiedAt",
                    "matchReasons",
                    "riskFlags",
                    "recommendedAction",
                    "postedAt",
                    "discoveredAt",
                    "lastSeenAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {}
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}